Created Locked Doors System
Created doorSO, keySO, DoorObject, KeyObject, and LockController as specified with Technical Design Document.
LockController tests DoorObject and KeyObject references to the identifiers in the given doorSO and keySO to see if the key matches the door as handled by the interaction manager.
Tester SOs and Game Objects in scene with debug logs within scripts to ensure working quality.