Created Clues
Created ClueController and ClueObject as per Technical Design Document. ClueObject contains reference to GoalSO associatedPuzzleStep. ClueController contains the current clue, the Action to when the clue is analyzed, and the enumerator which analyzes the clue in addition to it's public call function which sets the currentClue.
Tester functions, tester objects, and logging have all been kept for debugging purposes if necessary.