Camera Control v.1
Added: A new folder called Camera under the Scripts folder and a file called ThirdPersonCameraController.cs.
Added: Dummy capsule player with a target cube attached as a child, which is the object that the camera follows.
Added: The Cinemachine package to the project. A Cinemachine Brain component was attached to the main camera.
Added: A "Player" layer.